Begin work to punt sending data as well via the T2 timer.
This is a bit more than what JNOS does, so hold on.

The lapb_output() routine does a couple things:

* it sends what it can send, and
* if any data is sent, it clears axp->response which will clear any
  pending RR/RNR (but /not/ any pending REJs.)

However, the packet spam issue on transmit is due to ACK spam from
the sender.  Notably, every frame received by the stack will attempt
to call lapb_output(), which means if your sender decides to spam
you with RRs and REJs in weird ordering the sending stack may be
tricked into clearing the sent window tracking side and resend
stuff in its transmit window.

So, delaying lapb_output() to happen on t2 is an attempt at ensuring
that multiple back to back RR/REJ doesn't cause a lot of retransmission
of the same send stack over and over again.

This also means it's delayed by two seconds each time - unavoidable
if things are to be kept in sync correctly - but at least now you don't
get the stack deciding to burst out all kinds of weird nonsense.

Ok, so the change:

* create queue_lapb_output() which will queue LAPB output if required;
* axp->response is really now only used to communicate UA/disconnect
  through before disconnect.  The other places where it was being
  used is now using t2 and flags to send the relevant data.
* Only abort t2 in lapb_output() and sending an ACK if it's RR/RNR.
  If the stack decides the last thing it wants to send is a REJ
  then it won't be canceled and it'll go out once T2 fires.
* Send data before sending the ACK frame in send_ack().  yes, which
  really should be renamed.  That gives the stack the chance to
  not send an RR/RNR if it manages to send real data.

So far so good, but I haven't really sent a lot of bulk data through
this just yet.  I have some further testing to do before I'm comfortable
this is going to work out OK, and lord knows what other stacks will
do if I start not being so chatty towards them.